refactor(bootstrap): rename space_setup module to sync_engine
Phase 2 of the uc-bootstrap redesign. The space_setup module assembles
the entire iroh sync engine, yet its name both undersold that role and
collided with uc_application::facade::space_setup. Rename the module/file
to sync_engine to match what it builds.

Only the module path changes; the SyncEngineDeps parameter name
space_setup (a role label, not the module) and the unrelated
uc_application facade namesake are left untouched. The subsystem/ subdir
relocation is deferred to Phase 4.
